# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: BidSure Platform API
  version: 1.0.0
  description: |
    BidSure Platform API Contract — Phase 01 Architecture Baseline.
    
    ### System Conventions:
    - **API Prefix**: All endpoints are mounted under `/api/v1`.
    - **Request IDs**: Authoritative IDs generated by backend with format `req_<uuid>` (e.g., `req_550e8400-e29b-41d4-a716-446655440000`).
    - **Correlation IDs**: Optional client/system tracing identifier with format `corr_<uuid>`.
    - **Database IDs**: UUID v4 strings (frozen for Phase 03+).
    - **Timestamps**: Strictly ISO 8601 UTC format (`YYYY-MM-DDTHH:mm:ss.sssZ`).
    - **Envelope Standard**: All responses return standardized envelopes (`ApiResponseEnvelope` for 2xx, `ApiErrorEnvelope` for 4xx/5xx).
servers:
  - url: http://localhost:3000/api/v1
    description: Local Development Server
  - url: /api/v1
    description: Relative Path (Vite Dev Proxy)
paths:
...
  /bids/{bidId}/documents:
    get:
      summary: List Proposal Documents
      description: Returns all secure documents associated with a bid proposal.
      operationId: getBidDocuments
      parameters:
        - in: path
          name: bidId
          required: true
          schema:
            type: string
            format: uuid
          description: Bid proposal UUID
      responses:
        '200':
          description: List of documents attached to proposal.
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                  - requestId
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/BidDocumentDTO'
                  meta:
                    type: object
                    properties:
                      total:
                        type: integer
                  requestId:
                    type: string
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    post:
      summary: Upload Proposal Document
      description: Securely ingests, validates, hashes (SHA-256), scans, and stores a proposal document.
      operationId: uploadBidDocument
      parameters:
        - in: path
          name: bidId
          required: true
          schema:
            type: string
            format: uuid
          description: Bid proposal UUID
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - documentType
              properties:
                file:
                  type: string
                  format: binary
                  description: File payload (PDF, JPG, PNG; max 10MB)
                documentType:
                  type: string
                  enum: [TECHNICAL_BID, FINANCIAL_BID, COMPLIANCE_CERTIFICATE, EMD_PROOF, POWER_OF_ATTORNEY, OTHER]
                supersedesDocumentId:
                  type: string
                  format: uuid
                  description: Optional ID of document being replaced
      responses:
        '201':
          description: Document successfully ingested and enqueued for verification.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BidDocumentDetailResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /bids/{bidId}/documents/{documentId}:
    get:
      summary: Get Proposal Document Details
      description: Retrieves metadata and current ingestion status for a document.
      operationId: getBidDocumentById
      parameters:
        - in: path
          name: bidId
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: documentId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Document details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BidDocumentDetailResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      summary: Delete Proposal Document
      description: Unlinks and deletes a document from a DRAFT bid proposal. Prohibited on SUBMITTED bids.
      operationId: deleteBidDocument
      parameters:
        - in: path
          name: bidId
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: documentId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Document successfully removed.
        '409':
          description: Prohibited - proposal is submitted or finalized.
...
components:
  schemas:
...
    DocumentStatus:
      type: string
      enum:
        - UPLOADING
        - SCANNING
        - PROCESSING
        - READY
        - FAILED
        - QUARANTINED
        - REPLACEMENT_REQUIRED
      description: Authoritative document ingestion lifecycle state machine status.
    DocumentDTO:
      type: object
      required:
        - id
        - organizationId
        - fileName
        - originalFilename
        - mediaType
        - extension
        - sizeBytes
        - sha256
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        fileName:
          type: string
        originalFilename:
          type: string
        mediaType:
          type: string
          example: "application/pdf"
        extension:
          type: string
          example: "pdf"
        sizeBytes:
          type: integer
          example: 1048576
        sha256:
          type: string
          example: "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"
        status:
          $ref: '#/components/schemas/DocumentStatus'
        scanStatus:
          type: string
          nullable: true
          example: "PASSED"
        scanTimestamp:
          type: string
          format: date-time
          nullable: true
        failureCode:
          type: string
          nullable: true
        failureMessage:
          type: string
          nullable: true
        supersedesDocumentId:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    BidDocumentDTO:
      type: object
      required:
        - id
        - bidId
        - documentId
        - documentType
        - document
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        bidId:
          type: string
          format: uuid
        documentId:
          type: string
          format: uuid
        documentType:
          type: string
          enum: [TECHNICAL_BID, FINANCIAL_BID, COMPLIANCE_CERTIFICATE, EMD_PROOF, POWER_OF_ATTORNEY, OTHER]
        document:
          $ref: '#/components/schemas/DocumentDTO'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    BidDocumentDetailResponse:
      type: object
      required:
        - success
        - data
        - requestId
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/BidDocumentDTO'
        message:
          type: string
        requestId:
          type: string
          example: "req_550e8400-e29b-41d4-a716-446655440000"
...